initial user_chat, chat_message

+++ ska_backend/src/server/route_api/route_assistant/route_create_chat/dto_chat_details.rs
@@ -1,3 +1,4 @@
+use sea_orm::prelude::Decimal;
 use serde::{Deserialize, Serialize};
 
 #[derive(Clone, Serialize, Deserialize)]
@@ -6,8 +7,8 @@ pub struct DtoChatDetails {
     pub user_id: u32,
     pub llm_model: String,
     pub prompt_template: Option<String>,
-    pub temperature: Option<u32>,
-    pub top_p: Option<u32>,
+    pub temperature: Option<Decimal>,
+    pub top_p: Option<Decimal>,
     pub default_prompt: String,
 }
